Moved config into its own submodule.
