+++ copier/config/__init__.py
@@ -1,0 +1,1 @@
+from .factory import make_config, Flags  # noqa
